Implemented extends_included with & only args
We are now a (barly tested) feature mix of extends and include!
So far, extends_include seems like it mostly promotes bad practices.
Not that it adds anything bad, but it makes it easier not to seperate
subtemplates (such as a slightly modified generic form ) which extends would have
forced you to put in a seperate template. If there is even a slight chance that you will ever
need to use it twice, it realy should go in its own template... at least ATM.

This is still usefull in that context however, instead of using extends,
extends_included can be used. since extends_included can be used
multiple times in the same template, it can be usefull for building a
small template off of multiple generic templates.

Next experimental change will be exclude_blocks and only_blocks.
Consider the following
TEMPLATE_1:

```
{%block head-imports%}
<style>...custom>styles{...}...</style>
<script>...</script>
{%block head-imports%}
{%block body%}
{%for item in item_list%}...{%endfor%}
{%endblock body%}
```

TEMPLATE 2:

```
<head>
...OUR HEAD WHATEVER...
 # This is a concies and simple way to include a template, not requiring
 # you to seperate script/styling to remove redundency. and ensures only
 # one instance of <script> is included, and also alows you to place the
 # head-imports exactly where you want it (after jQuery, but before <styles> for instance).

{% include "TEMPLATE_1" only_blocks "head-imports" %}
</head>
<body>
{% include "TEMPLATE_1" exclude_blocks "head-imports" with item_list=list1 %}
{% include "TEMPLATE_1" exclude_blocks "head-imports" with item_list=list2 %}
{% include "TEMPLATE_1" exclude_blocks "head-imports" with item_list=list3 %}
</body>
```